docs(autosave): split the recipe into its own file, add a per-field draft gate
Make the autosave demo a folder demo: useAutosave moves into its own
useAutosave.ts that App.vue imports, so the playground shows the
composable and its consumer as separate pages and the file organization
mirrors how you'd drop it into an app.

The recipe's gateOnValidity now also takes a per-path predicate, and the
demo uses it to keep email a draft (ungated, .email()-validated) while
display name and bio gate on validity. Typing an invalid email still
autosaves the draft while form.fields.email.showErrors surfaces the
validation message: the write persists, the read flags it, neither
blocks the other. autosave.md is synced to match.
